LevelSets

Documentation for LevelSets.

LevelSets.LevelSets

LevelSets.LevelSet

LevelSets.boundlevelset

LevelSets.boundlevelsetgrad!

LevelSets.difflevelset!

LevelSets.difflevelset!

LevelSets.difflocallevelset!

LevelSets.difflocallevelset!

LevelSets.evallevelset

LevelSets.findclosest!

LevelSets.hessianlevelset!

LevelSets.hesslocallevelset!

LevelSets.jacobian!

LevelSets.locallevelset

LevelSets.parameterindices

LevelSets.residual!

LevelSets.LevelSets — Module

Implements smooth-max inspired level-set functions.

LevelSets.LevelSet — Type

Data required by smooth-max level-set function.

LevelSets.boundlevelset — Method

ls, bound = boundlevelset!(x, dx, levset)

Return level-set value and bound defined by levset, over the element centered at x and with half side lengths dx.

LevelSets.boundlevelsetgrad! — Method

bound = boundlevelsetgrad!(x_bar, x, dx, levset)

Return bound on level-set gradient x_bar, defined by levset, over the element centered at x and with half side lengths dx.

LevelSets.difflevelset! — Method

difflevelset!(xcenter_bar, frame_bar, kappa_bar, rho_bar, x, levset)

Compute the derivatives of the level set, defined by levset, at point x with respect to the LevelSet parameters.

Uses the reverse-mode of algorithmic differentiation (i.e. back propagation).

LevelSets.difflevelset! — Method

difflevelset!(x_bar, x, levset)

Compute the derivatives of the level set, defined by levset, at point x with respect to x. The derivatives are stored in x_bar.

Uses the reverse-mode of algorithmic differentiation (i.e. back propagation).

LevelSets.difflocallevelset! — Method

difflocallevelset!(xc_bar, frame_bar, kappa_bar, x, xc, frame, kappa, perp_bar)

Computes the dervatives of locallevelset with respect to xc, frame, and kappa. The derivatives are returned in the _bar variables and weighted by perp_bar.

LevelSets.difflocallevelset! — Method

difflocallevelset!(x_bar, x, xc, frame, kappa, perp_bar)

Computes the dervatives of locallevelset with respect to x. The derivatives are returned in x_bar and are weighted by perp_bar

LevelSets.evallevelset — Method

ls = evallevelset(x, levset)

Return the level-set value at x defined by levset.

LevelSets.findclosest! — Method

findclosest!(x, x0, levset[, tol=1e-12, max_newton=10])

Compute the point x that is on the zero level contour and closest to x0.

LevelSets.hessianlevelset! — Method

hessianlevelset!(hess, x, levset)

Compute the Hessian of the level set, defined by levset, at point x with respect to x. The derivatives are stored in hess.

LevelSets.hesslocallevelset! — Method

hesslocallevelset!(hess, x, xc, frame, kappa)

Computes the Hessian of locallevelset with repsect to x. Note that xc and x are provided but not actually used; quadatic level-set is constant. We'll keep this interface in case higher-order local level-sets are used in the future.

LevelSets.jacobian! — Method

jacobian!(jac, x, levset)

Evaluate the Jacobian of residual! at x with respect to levset parameters.

LevelSets.locallevelset — Method

ls_local = locallevelset(x, xc, frame, kappa)

Return a quadratic approximation to the level set based on given xc, frame coordinate axes, and curvatures kappa.

LevelSets.parameterindices — Method

xc_idx, nrm_idx, len_idx = parameterindices(levset)

Create indices for xcenter, normal, and lenscale parameters in levset.

LevelSets.residual! — Method

residual!(res, x, levset)

Evaluate the level-set defined by levset at points x, and store in res