LevelSets

Documentation for LevelSets.

LevelSets.LevelSets

LevelSets.LevelSet

LevelSets.diffexpdist!

LevelSets.difflevelset!

LevelSets.difflevelset!

LevelSets.difflocallevelset!

LevelSets.difflocallevelset!

LevelSets.evallevelset

LevelSets.expdist

LevelSets.hessexpdist!

LevelSets.hessianlevelset!

LevelSets.hesslocallevelset!

LevelSets.locallevelset

LevelSets.residual!

LevelSets.snappoint!

LevelSets.LevelSets — Module

Implements smooth-max inspired level-set functions.

LevelSets.LevelSet — Type

Data required by smooth-max level-set function.

LevelSets.diffexpdist! — Method

diffexpdist!(x_bar, x, xc, rho, delta, min_dist, exp_bar)

Computes the derivative of expdist with respect to x and returns the result in the array x_bar. The derivative is weighted by exp_bar.

LevelSets.difflevelset! — Method

difflevelset!(xcenter_bar, frame_bar, kappa_bar, rho_bar, x, levset)

Compute the derivatives of the level set, defined by levset, at point x with respect to the LevelSet parameters.

Uses the reverse-mode of algorithmic differentiation (i.e. back propagation).

LevelSets.difflevelset! — Method

difflevelset!(x_bar, x, levset)

Compute the derivatives of the level set, defined by levset, at point x with respect to x. The derivatives are stored in x_bar.

Uses the reverse-mode of algorithmic differentiation (i.e. back propagation).

LevelSets.difflocallevelset! — Method

difflocallevelset!(xc_bar, frame_bar, kappa_bar, x, xc, frame, kappa, perp_bar)

Computes the dervatives of locallevelset with respect to xc, frame, and kappa. The derivatives are returned in the _bar variables and weighted by perp_bar.

LevelSets.difflocallevelset! — Method

difflocallevelset!(x_bar, x, xc, frame, kappa, perp_bar)

Computes the dervatives of locallevelset with respect to x. The derivatives are returned in x_bar and are weighted by perp_bar

LevelSets.evallevelset — Method

ls = evallevelset(x, levset)

Return the level-set value at x defined by levset.

LevelSets.expdist — Method

expfac = expdist(x, xc, rho, delta, min_dist)

Computes the exponential of the scaled distance from x to xc. The distance is regularized by adding delta to the dot product of x - xc; this helps with derivatives when norm(x - xc) is small. The scalar -rho scales the difference between this regularized distance and some reference min_dist.

LevelSets.hessexpdist! — Method

hessexpdist!(hess, x, xc, rho, delta, min_dist)

Computes the Hessian of expdist with respect to x and returns the result in the array hess.

LevelSets.hessianlevelset! — Method

hessianlevelset!(hess, x, levset)

Compute the Hessian of the level set, defined by levset, at point x with respect to x. The derivatives are stored in hess.

LevelSets.hesslocallevelset! — Method

hesslocallevelset!(hess, x, xc, frame, kappa)

Computes the Hessian of locallevelset with repsect to x. Note that xc and x are provided but not actually used; quadatic level-set is constant. We'll keep this interface in case higher-order local level-sets are used in the future.

LevelSets.locallevelset — Method

ls_local = locallevelset(x, xc, frame, kappa)

Return a quadratic approximation to the level set based on given xc, frame coordinate axes, and curvatures kappa.

LevelSets.residual! — Method

residual!(res, x, levset)

Evaluate the level-set defined by levset at points x, and store in res

LevelSets.snappoint! — Method

snappoint!(x, x0, levset[, tol=1e-12, max_newton=20])

Find a point on the level set that is "close" to x0. The solution, x, satisfies |phi(x)| < tol. However, x may not be the point on the level set that is closest to x0.